Move app enrichment of user sessions to users SDK
