+++ packages/server/src/sdk/users/index.ts
@@ -1,5 +1,7 @@
 import * as utils from "./utils"
+import * as sessions from "./sessions"
 
 export default {
   ...utils,
+  sessions,
 }
